refactor(agentcore): leave the process-global sink with the process entry
